Changes in migration number 10 wrong table name and references

+++ src/main/resources/db/migration/V10__create_specialization_tables.sql
@@ -15,7 +15,7 @@ CREATE TABLE specialization (
 -- Create doctor_specialization_mapping table
 CREATE TABLE doctor_specialization_mapping (
     ds_id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    doctor_id BIGINT NOT NULL,
+    doctor_id INT NOT NULL,
     specialization_id BIGINT NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     created_by BIGINT,
